Generalize log_exception helper
