+++ linkding_cli/helpers/logging.py
@@ -1,7 +1,12 @@
 """Define logging helpers."""
 from __future__ import annotations
 
+from functools import wraps
+from typing import Any, Callable, TypeVar
+
 import typer
+
+T = TypeVar("T")
 
 
 def error(msg: str) -> None:
@@ -14,3 +19,25 @@ def debug(ctx: typer.Context, msg: str) -> None:
     if not ctx.obj.config.verbose:
         return
     typer.echo(f"Debug: {msg}")
+
+
+def log_exception(
+    *, exit_code: int = 1
+) -> Callable[[Callable[..., T]], Callable[..., T]]:
+    """Define a dectorator to handle exceptions via typer output."""
+
+    def decorator(func: Callable[..., T]) -> Callable[..., T]:
+        """Decorate."""
+
+        @wraps(func)
+        def wrapper(*args: Any, **kwargs: dict[str, Any]) -> T:
+            """Wrap."""
+            try:
+                return func(*args, **kwargs)
+            except Exception as err:  # pylint: disable=broad-except
+                error(str(err))
+                raise typer.Exit(code=exit_code) from err
+
+        return wrapper
+
+    return decorator
